Move dependency on ContainerService for models to ContainerModelService
