+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/store/SequencingContainerModelStore.java
@@ -29,7 +29,7 @@ import uk.ac.bbsrc.tgac.miso.core.data.Platform;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.SequencingContainerModel;
 
 public interface SequencingContainerModelStore {
-  SequencingContainerModel getModel(long id);
+  SequencingContainerModel get(long id);
 
   /**
    * Attempt to find a SequencingContainerModel matching the supplied parameters
@@ -42,8 +42,8 @@ public interface SequencingContainerModelStore {
    *          the number of partitions that the model must have (required)
    * @return an appropriate model if one is found; null otherwise
    */
-  SequencingContainerModel findModel(Platform platform, String search, int partitionCount);
+  SequencingContainerModel find(Platform platform, String search, int partitionCount);
 
-  List<SequencingContainerModel> listModels();
+  List<SequencingContainerModel> list();
 
 }
